add corn job

+++ database/migrations/2022_10_09_185533_create_order_table.php
@@ -18,7 +18,7 @@ class CreateOrderTable extends Migration
             $table->string("name");
             $table->string("email");
             $table->string("empy_id");
-            $table->integer("phone");
+            $table->bigInteger("phone");
             $table->text("des");
             $table->string("title");
             $table->string("receipt");
